+++ src/Ext/Techno/Hooks.cpp
@@ -46,6 +46,17 @@ DEFINE_HOOK(0x6F9E50, TechnoClass_AI, 0x5)
 
 	}
 
+	if (pExt->DelayedFire_Anim && pThis->GetCurrentMission() != Mission::Attack)
+	{
+		CDTimerClass* weaponReadyness = (CDTimerClass*)pThis->__DiskLaserTimer;
+		weaponReadyness->Start(pThis->DiskLaserTimer.GetTimeLeft() + 5);
+
+		// Reset Delayed fire animation
+		pExt->DelayedFire_Anim = nullptr;
+		pExt->DelayedFire_Anim_LoopCount = 0;
+		pExt->DelayedFire_DurationTimer = -1;
+	}
+
 	return 0;
 }